requirements.txt
Django>=5.1.9
djangorestframework>=3.14.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
